Add `evm_coinbase`
Adds a non-standard RPC method to set the coinbase for mined blocks.

NOTE: Unless this is called, the value returned by `eth_coinbase`
actually differs from the coinbase included in blocks. The default for
mined blocks is actually the null address, and I opted not to change it
to not break backwards compatibility.