Fix broken --host option due to being overridden by help option
The help option was defined to use the 'h' shorthand option.
However this was overriding the previous host option which also
used the 'h' shorthand option. Thus running ganache-cli with the
'-h' or even '--host' option did not work and showed the help output.
Now only '--help' and '-?' trigger the help output.